requirements.txt
Django==1.11
psycopg2==2.7.1
wheel==0.24.0
django-bootstrap3==8.2.3
django-crispy-forms==1.6.1
django-filter==1.0.4
django-model-utils==3.0.0
django-rest-swagger==2.1.2
djangorestframework==3.6.3
djangorestframework-jwt==1.10.0
Markdown==2.6.8
whitenoise==3.0
django-crontab==0.7.1

